+++ src/gradient/resources/knowledge_bases/indexing_jobs.py
@@ -266,8 +266,8 @@ class IndexingJobsResource(SyncAPIResource):
         self,
         uuid: str,
         *,
-        poll_interval: int = 5,
-        timeout: int | None = None,
+        poll_interval: float = 5,
+        timeout: float | None = None,
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
         # The extra values given here take precedence over values defined on the client or passed to this method.
         extra_headers: Headers | None = None,
@@ -601,8 +601,8 @@ class AsyncIndexingJobsResource(AsyncAPIResource):
         self,
         uuid: str,
         *,
-        poll_interval: int = 5,
-        timeout: int | None = None,
+        poll_interval: float = 5,
+        timeout: float | None = None,
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
         # The extra values given here take precedence over values defined on the client or passed to this method.
         extra_headers: Headers | None = None,
